moved screenrecord function to a different file
